--- .firebase.js
@@ -1,9 +1,0 @@
-const firebaseConfig = {
-  apiKey: "AIzaSyAVv-8TSjgDbnq3C1Oso6mOWjFB0eF8-cw",
-  authDomain: "react-chatbot-2e80c.firebaseapp.com",
-  databaseURL: "https://react-chatbot-2e80c.firebaseio.com",
-  projectId: "react-chatbot-2e80c",
-  storageBucket: "react-chatbot-2e80c.appspot.com",
-  messagingSenderId: "999819616337",
-  appId: "1:999819616337:web:d9121bff04bef9e4f11870"
-};
